Compiles and runs under VS2008. Ues boost::sub_range.

+++ src/ai/defaultai.cc
@@ -958,7 +958,7 @@ bool DefaultAI::construct_building (int32_t) // (int32_t gametime)
 				if (bf->water_nearby < 3)
 					continue;
 				int effect = bf->water_nearby - 8;
-				prio += effect > 0 ? static_cast<int>(sqrt(effect)) : effect;
+				prio += effect > 0 ? static_cast<int>(sqrt((double)effect)) : effect;
 				// if same producers are nearby, then give some penalty
 				for (size_t k = 0; k < bo.outputs.size(); ++k)
 					if (bf->producers_nearby[bo.outputs[k]] > 0)
@@ -1790,7 +1790,7 @@ int32_t DefaultAI::calculate_need_for_ps(BuildingObserver & bo, int32_t prio)
 
 	if (bo.outputs.size() > 1)
 		output_prio = static_cast<int32_t>
-			(ceil(output_prio / sqrt(bo.outputs.size())));
+			(ceil(output_prio / sqrt((double)bo.outputs.size())));
 	prio += output_prio;
 
 	// If building consumes some wares, multiply with current statistics of all
